Corrected the Id type for the repository

+++ src/main/java/guru/springframework/repositories/DatabaseRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import guru.springframework.domain.LabDatabase;
 
 
-public interface DatabaseRepository extends PagingAndSortingRepository<LabDatabase, Integer>{
+public interface DatabaseRepository extends PagingAndSortingRepository<LabDatabase, Long>{
 
 }
